articleで行う処理を別ファイルに分割し、 application.jsで読み込みをした
articleではいいね画像に対するget, post, deleteを実装
